+++ lib/flutter_dsl.dart
@@ -1,6 +1,6 @@
 library flutter_dsl;
 
-export 'src/widget_extensions.dart';
-export 'src/text_extensions.dart';
-export 'src/spacing_extensions.dart';
 export 'src/conditional_extensions.dart';
+export 'src/spacing_extensions.dart';
+export 'src/text_extensions.dart';
+export 'src/widget_extensions.dart';